Add separate optimization phase on core
